Added delete functionality
I realized something while doing delete. In insert, I used a recursive call to handle deleting elements from a reversed list in the case of adding elements between the ends of the list, where I recalculate the position using the formula "non-reversed position = size - position - 1." I then recursively called insert with the non-reversed position.

I realized while writing del that I could have just used a single else statement for deleting elements that are not at either end of the list, check if the list is reversed, if so recalculate the position to reflect that of the non-reversed list, and then proceed as if the list was not reversed, circumventing the need for a recursive call. The same principle should apply to insert. The recursion, while clever at the time, was unnecessary and I should correct it to avoid having the added complexity and wasting resources on the stack (while probably minor in this case, it's important not to get carried away and try to find alternative methods that do not add needless complexity or waste resources. The index recalculation was a good step that allows me to achieve constant time list reversibility, which is a plus)